requirements.txt
SpeechRecognition
pyttsx3
wikipedia
pywhatkit
pyaudio
requests
pyjokes
python-dotenv